feat: Extract markdown component into separate file
