+++ app/src/app/copilot/chat/bubble.tsx
@@ -2,8 +2,7 @@
 
 import { type Message } from "ai";
 import { cn } from "~/lib/utils";
-import { CodeBlock } from "../markdown/code-block";
-import { MemoizedReactMarkdown } from "../markdown/react-markdown";
+import { Markdown } from "../markdown/react-markdown";
 import { Avatar } from "~/components/ui/avatar";
 import { RateMessage } from "./rate-message";
 
@@ -133,92 +132,7 @@ export default function Bubble({
         )}
 
         <div className="mt-1 w-full space-y-3">
-          <MemoizedReactMarkdown
-            className="prose flex-1 dark:prose-invert"
-            // remarkPlugins={[remarkGfm, remarkMath]}
-            // // rehypePlugins={[rehypeMathjax]}
-            components={{
-              code({ className, children, ...props }) {
-                if (
-                  Array.isArray(children) &&
-                  typeof children[0] === "string"
-                ) {
-                  if (children[0] === "▍") {
-                    return (
-                      <span className="mt-1 animate-pulse cursor-default">
-                        ▍
-                      </span>
-                    );
-                  }
-
-                  children[0] = children[0].replace("`▍`", "▍");
-                }
-
-                const match = /language-(\w+)/.exec(className ?? "");
-
-                return "inline" in props && !props.inline ? (
-                  <CodeBlock
-                    key={Math.random()}
-                    language={match?.[1] ?? ""}
-                    value={String(children).replace(/\n$/, "")}
-                    {...props}
-                  />
-                ) : (
-                  // Apply the inline-code class along with any existing classes
-                  <code className={cn("inline-code", className)} {...props}>
-                    {children}
-                  </code>
-                );
-              },
-              table({ children }) {
-                return (
-                  <table className="border-collapse border border-black px-3 py-1 dark:border-white">
-                    {children}
-                  </table>
-                );
-              },
-              th({ children }) {
-                return (
-                  <th className="break-words border border-black bg-gray-500 px-3 py-1 text-white dark:border-white">
-                    {children}
-                  </th>
-                );
-              },
-              td({ children }) {
-                return (
-                  <td className="break-words border border-black px-3 py-1 dark:border-white">
-                    {children}
-                  </td>
-                );
-              },
-              a({ href, children, ...props }) {
-                // Check if the last child is a string and if it ends with a period
-                if (
-                  Array.isArray(children) &&
-                  typeof children.at(-1) === "string"
-                ) {
-                  const lastChild = children.at(-1) as string;
-
-                  return (
-                    <>
-                      <a
-                        href={href}
-                        className="markdown-link text-blue-400"
-                        {...props}
-                      >
-                        {children}
-                      </a>
-                      {!lastChild.trim().endsWith(".") && ". "}
-                    </>
-                  );
-                }
-
-                return <>{children}</>;
-              },
-            }}
-          >
-            {`${messageWithOutSources}`}
-          </MemoizedReactMarkdown>
+          <Markdown>{`${messageWithOutSources}`}</Markdown>
 
           <div className="flex w-full items-center justify-between">
             <div className="flex w-full items-center space-x-3">
